Moved physType physics to new functions
