+++ src/index.js
@@ -9,7 +9,7 @@ import './custom.css';
 
 // our own stuff
 import { actionDispatch, startSim, stopSim } from './reduxlike/action_creators.js';
-import { doUpdateLoop } from './do_update_loop.js';
+import { doUpdateLoop } from './sim/do_update_loop.js';
 import { storeInit } from './reduxlike/store_init.js';
 import { RULE_HIT_WALL, RULE_CONDS_OUT_OF_LIMITS } from './rulebook.js';
